Remove /Zc:preprocessor from Windows preset
While this flag would be very nice to have to weed out MSVC extensions
from the preprocessro and have standards conforming code, it requires a
certain Windows SDK version, which is not the default on GitHub Actions
windows image and likely would be bothersome to set up locally on the
developer's machine as well.

To make things simpler, we rely on the *nix CI runs to verify the
standards conformance of the code.