Add --vcpkg flag for vcpkg port generation
The --vcpkg flag is a "mode" flag similar to how cmake --build and cmake
--install are mode flags for CMake.

This flag will generate port files that are designed to easily integrate
with projects also generated by cmake-init, but the basic structure of
the generated vcpkg port file is very similar to ones written for other
projects.

This vcpkg integration allows users to pull in any dependency from
GitHub with minimal effort, however this requires putting the entire
dependency graph into the version control in the form of port files.
This isn't ideal, but this is still the option that requires the least
amount of effort to reuse projects from GitHub without submitting them
to a central package repository.