Move examples from dev only to the general area
Consumers of the libraries might want to build the examples themselves
to play around with, so it's better to them to the general area of the
CML.

It's still guarded by an option whose value depends on the dev mode by
default, i.e. they are enabled the first time the project is configured
in dev mode.