Add clang-format support
Introduce the lint.cmake script, which is used both by CMake build
targets to lint and fix the code in developer mode and by the CI to
check the source code format before any other CI job.

The `format-check` and `format-fix` targets can be used to check and
fix the source code respectively using clang-format.