Revert changes in windows-set-path.cmake
Passing the PATH to the test executable is not possible, because the
test discovery modules all make a call to the test executable where the
developer has no control over passing environment variables to the test
executable.

On Windows, this will cause errors, because the executable will not be
able to locate its shared library dependencies.