Speed up header-only template
Specifying no language in the project call prevents compiler detection,
which can take a while on Windows systems.

When this is done and GNUInstallDirs is included, then the
CMAKE_INSTALL_LIBDIR path cache variable must be set by the project code
to a sensible default (i.e. `lib`), because having no compiler detected
also causes the void pointer's size to be left unknown to CMake and
GNUInstallDirsdepends on its value when setting up the
CMAKE_INSTALL_LIBDIR variable.

Setting CMAKE_INSTALL_LIBDIR before the inclusion of GNUInstallDirs will
cause the variable to be processed the same as if the module defined it.