Enable building docs without configuring
Configuring the project can take a while, so to avoid doing that in CI,
the docs.cmake script was modified to be able to run in script mode.

The downside of this is that you may not use variables from the lists
file that are defined after the project() call, because the script
overrides the project() command to set its arguments as variables, then
stop execution of the lists file. This is to avoid "parsing" the
project() call with regexes, which are always painfully fragile.

Since the docs.cmake script now runs in script mode, m.css is not
fetched with FetchContent, but with manual command calls. This is
because FetchContent cannot run in script mode.