Move the dev mode code to a script
The dev mode code at the end of each lists file was getting long, but
also a modification to it required changes across multiple files.

This part was originally not extracted into a script, because the
examples used to be here and managing that was just something I didn't
want to deal with, but now that the examples are part of the main code
path, this is not relevant.