Simplify docs generation in CI
docs.cmake was cut down in size due to not having to handle being run as
either a script or as a module. Documentation generation is now done by
configuring the project, however there is a function that docs.cmake
overrides if included with project() code injection which will return
early in the config process. This saves time and resources when the
project would otherwise require downloading and installing dependencies.

There is an additional opportunity to reduce the time it takes to
configure in CI, which is moving the enabling of languages from the
project() call after the docs_early_return() call, but this would
deviate from the usual structure of lists files too much and so it's not
deemed worth using. CI uses Ubuntu to run the docs job, which means that
configuring a project is very fast either way.