Make include quoting consistent
Headers coming from outside the project are included with with angle
brackets (<>), while project specific headers are included with double
quotes. There is a little trade-off to be made here, because in
header-only and static/shared library mode the tests can also be run by
importing the main project as an outside dependency, but since the CI is
setup in a way that this scenario is not tested there, this is an easy
compromise to make.