Attempt to improve SEO
Right now the search engines pick up the description for the result from
the License section of the readme, which I can only assume to be,
because it's the only paragraph that starts with "cmake-init". Beginning
the introductory description with "cmake-init" will hopefully provide a
better description for search engines to use.