Apply the previous commits to `run_examples` too
The delimiter for the global utility target was changed to a hyphen,
however the individual run targets that are created with the add_example
command call are left with an underscore as a delimiter, because their
names are derived from files on the filesystem, which will most likely
use underscores for delimiters as well. These individual targets however
are not part of the "tips" message printed by cmake-init, so on the
surface level every utility target appears to use hyphens.