Add a new preset with the standard enforced
The CMAKE_*_EXTENSIONS variable has an effect only when used in
combination with the CMAKE_*_STANDARD variable, and not in combination
with the compile features.

Having these variables moved into a separate ci-std preset makes it
explicit that these are only for testing in the CI and making sure the
project actually builds with the specified standard.