Move C4251 suppression entirely into the header
The macro function should be populated in the header itself, because not
only are there more compilers on Windows than just MSVC, but this header
gets installed, so the check needs to happen in the consuming code.

The check for __ICL is taken from the Hedley library, which makes this
suppression exclude the Intel compiler on Windows. I don't have the
Intel compiler to check whether this is correct.