Use presets version 2
Presets version 2 requires CMake 3.20.

This commit allows users with VSCode to comfortably use the generated
projects with the "CMake Tools" extension out of the box.

Hints printed after project generation and generated documentation have
been adjusted to use the new capabilities of the presets, which results
in simpler workflow for developers regardless of what build system or
operating system they use.

The CI workflows or the checked in presets file have not been further
adjusted, because the additional build and test presets aren't as useful
there. The workflows are already deduplicated and serve as a sort of
preset on their own. Users may add their own if that better suits their
tastes.

Fixes #34