Add usage info to BUILDING.md
This usage info is only generated for library projects, because it turns
out that getting the executable path is a bit tricky when considering
that multi-config generators will still install executable targets with
their specific config properties which were used to compile it.

The above means that an executable built in debug mode will be installed
with the IMPORTED_LOCATION_DEBUG property, so getting the path is not as
straightforward.

I just decided to omit info for usage within CMake for installed
executable projects. I'll figure out later how to properly add an
XXX_EXECUTABLE variable like many of the built-in modules from CMake.