Add a separate docs script for CI
This script will fetch m.css on its own and parse the project() call out
of the top level CML for its arguments.

This solution is a bit messy still. One idea to improve this would be to
create the docs artifacts in one of the CI runs and then use those in
the publishing job. This falls apart when the publishing is run on the
master branch only, like one of the example cmake-init projects. There,
the testing is done on the develop branch and the GH pages is populated
with docs generated on pushes to the master branch.

Kind of related issue: #41
This space could use improvements.