Implement a very simple template engine
This template engine renders files in a single pass and allows for
nested control flow in the templates. It's similar to Jinja2, except
extremely barebones and there is no pretty error reporting for mistakes
in the templates. The only error that is reported early is when blocks
weren't terminated properly. This check is done using a depth counter,
so don't expect too much.