Inject the template compiler via `main` parameter
This is very hacky. When packaged as a PyPI wheel, the module path to
the template.py file is ".template", but when importing in a pyz file,
just "template" works.

Now, the importing of the template file is done at the entrypoint and is
injected into the main function of the cmake_init.py file.