Use all conformance flags for VS2022 preset
The externConstexpr flag was simply something I overlooked in the past.

Since the project now assumes VS2022, the preprocessor flag can be
safely used.
Note that old Windows SDKs will cause warnings to be emitted regardless,
but assuming that the user will use a new or the latest SDK when using
VS2022 is not unreasonable.