Use conanfile.py to simplify install command
Using the Python script allows setting the output folder to "conan",
which the CMake presets expect to be populated by the time CMake is
invoked. This simplifies the command which one has to execute as a
developer and makes the generated projects compatible with Conan v2,
since that version removes the "-if" flag for the install command.