Put the CMake package in `lib` for libraries
Previously, the CMake package for both library and header-only porjects
went in the "share" directory, however it's more appropriate to put the
CMake package in the architecture specific "lib" directory on platforms
where it applies.

The "share" directory is for architecture agnostic read-only data, but
the CMake package of a library project is tied to the compiled binary,
which is not architecture agnostic. Thus, it's better to use
${CMAKE_INSTALL_LIBDIR}/cmake/<package> for the CMake package.