Clarify the licensing of the templates
The contents of the cmake-init/templates directory were previously
claimed to be "Public Domain" in the README, however this public domain
does not exist in some jurisdictions like some parts of Europe. Using
the Unlicense license makes sure that my intent applies in every
situation so that the output of cmake-init can be used by anyone for any
purpose.