Simplify use of `json_object_get_string_len`
This function will always return a value in the range [0, INT_MAX] and
casting that value to size_t for malloc will always result in a safe
operation once incremented to account for the null terminator.

You could argue that accounting for the null terminator is pointless if
all this function returns is a char* to the name of the project, but I
would prefer to correctly use the dependency in the example code.