Use package's found variable in the config script
Since "find_dependency" sets the found variable to false, this can be
used to conveniently check at the end before including the export set
whether all dependencies were found in a non-REQUIRED call.

This requires setting the <name>_FOUND variable to true at the start of
the file, because "find_dependency" sets the variable only if something
could not be found.